+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+- fixed `ConsulWatch` in xresolver by storing and watching the correct part of the url
 
 ## [v1.10.1]
 ### Fixed